shiny network and line graph
Updated the line graph to allow for many options. I had to use an example provided on stack overflow for a javascript function that allows me to enter many player options. I also worked for a long time to create a network graph, but I am a little unsatisfied with the final result as it is a little too busy.